extracted colors to trs-system and implemented some more custom uis

+++ oscar/preferences/Defaults.swift
@@ -10,6 +10,6 @@ import Defaults
 extension Defaults.Keys {
     static let sampleFreq = Key<Double>("sample_freq", default: 100)
     static let maxSamples = Key<Int>("max_samples", default: 100)
-    static let defaultPort = Key<UInt16>("default_port", default: 8_000)
+    static let defaultPort = Key<Int>("default_port", default: 8_000)
     static let isShowingDebugConsole = Key<Bool>("is_showing_debug_console", default: false)
 }
